requirements.txt
antiorm==1.2.1
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==41.0.0
db==0.1.1
db-sqlite3==0.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.10
django-cors-headers==4.0.0
django-phonenumber-field==7.1.0
django-rest-framework-nested==0.0.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
drf-yasg==1.21.5
dynaconf==3.1.12
hvac==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
knox==0.1.14
loguru==0.7.0
Markdown==3.4.3
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
phonenumbers==8.13.13
Pillow==9.5.0
pycodestyle==2.10.0
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.7.0
pyrsistent==0.19.3
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.29
ruamel.yaml.clib==0.2.7
shortuuid==1.0.11
simplejson==3.19.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.2
validators==0.20.0
dj-rest-auth